requirements.txt
opencv-python==4.5.5.64
numpy==1.19.2
streamlit==1.9.0
json5==0.9.5
Pillow==8.0.1

